refactor: move runtime logs message in details package
