+++ gravitee-apim-console-webui/src/management/api/runtime-logs-v4/runtime-logs-details/runtime-logs-messages/api-runtime-logs-messages.component.scss
@@ -1,7 +1,7 @@
 @use 'sass:map';
 @use '@angular/material' as mat;
 @use '@gravitee/ui-particles-angular' as gio;
-@use '../../../../scss/gio-layout' as gio-layout;
+@use 'src/scss/gio-layout' as gio-layout;
 
 $typography: map.get(gio.$mat-theme, typography);
