+++ gravitee-apim-console-webui/src/management/api/runtime-logs-v4/runtime-logs-details/runtime-logs-messages/components/api-runtime-logs-message-connection-log/api-runtime-logs-message-connection-log.component.scss
@@ -1,6 +1,6 @@
 @use 'sass:map';
-@use '@angular/material' as mat;
-@use '@gravitee/ui-particles-angular' as gio;
+@use '../../../../../../../../node_modules/@angular/material/index' as mat;
+@use '../../../../../../../../node_modules/@gravitee/ui-particles-angular/index' as gio;
 
 $typography: map.get(gio.$mat-theme, typography);
 
@@ -37,16 +37,13 @@ $typography: map.get(gio.$mat-theme, typography);
     margin-top: 8px;
     margin-bottom: 8px;
 
-    &:has(.gio-badge-primary) {
-      margin-top: 24px;
-      margin-bottom: 0;
-    }
-
-    &:has(api-runtime-logs-message-item-content) {
-      margin-top: 0;
-    }
-
     &__cell {
+      display: flex;
+      flex-direction: row;
+    }
+
+    &__sub-cell {
+      margin: 8px 0 8px 12px;
       display: flex;
       flex-direction: row;
     }
@@ -60,24 +57,7 @@ $typography: map.get(gio.$mat-theme, typography);
 
     &__value {
       word-break: break-all;
-      color: mat.get-color-from-palette(gio.$mat-space-palette, 'default');
       @include mat.typography-level($typography, body-1);
-    }
-
-    &__value.gio-badge-warning,
-    &__value.gio-badge-accent {
-      @include mat.typography-level($typography, caption);
-    }
-  }
-
-  .gio-badge-primary,
-  .gio-badge-neutral,
-  .gio-badge-accent,
-  .gio-badge-warning {
-    font-weight: 700;
-
-    mat-icon {
-      margin-right: 4px;
     }
   }
 }
