+++ gravitee-apim-console-webui/src/management/api/runtime-logs-v4/runtime-logs-details/runtime-logs-messages/components/api-runtime-logs-message-empty/api-runtime-logs-message-empty.component.scss
@@ -1,6 +1,6 @@
 @use 'sass:map';
-@use '@angular/material' as mat;
-@use '@gravitee/ui-particles-angular' as gio;
+@use '../../../../../../../../node_modules/@angular/material/index' as mat;
+@use '../../../../../../../../node_modules/@gravitee/ui-particles-angular/index' as gio;
 
 .empty-card {
   text-align: center;
