+++ gravitee-apim-console-webui/src/management/api/runtime-logs-v4/runtime-logs-details/runtime-logs-messages/components/api-runtime-logs-message-item/api-runtime-logs-message-item-content.component.ts
@@ -17,7 +17,7 @@
 import { Component, Input } from '@angular/core';
 import { FormControl } from '@angular/forms';
 
-import { Message } from '../../../../../../entities/management-api-v2';
+import { Message } from '../../../../../../../entities/management-api-v2';
 
 type ContentTab = {
   label: string;
