+++ gravitee-apim-console-webui/src/management/api/runtime-logs-v4/runtime-logs-details/runtime-logs-messages/components/api-runtime-logs-message-item/api-runtime-logs-message-item.component.ts
@@ -16,7 +16,7 @@
 
 import { Component, Input } from '@angular/core';
 
-import { AggregatedMessageLog } from '../../../../../../entities/management-api-v2';
+import { AggregatedMessageLog } from '../../../../../../../entities/management-api-v2';
 
 @Component({
   selector: 'api-runtime-logs-message-item',
