+++ gravitee-apim-console-webui/src/management/api/runtime-logs-v4/runtime-logs-details/runtime-logs-messages/components/api-runtime-logs-message-item/api-runtime-logs-message-item.stories.ts
@@ -19,10 +19,10 @@ import { Story } from '@storybook/angular/dist/ts3.9/client/preview/types-7-0';
 import { ApiRuntimeLogsMessageItemModule } from './api-runtime-logs-message-item.module';
 import { ApiRuntimeLogsMessageItemComponent } from './api-runtime-logs-message-item.component';
 
-import { fakeAggregatedMessageLog, fakeMessage } from '../../../../../../entities/management-api-v2';
+import { fakeAggregatedMessageLog, fakeMessage } from '../../../../../../../entities/management-api-v2';
 
 export default {
-  title: 'API / Logs / Messages / Item',
+  title: 'API / Logs / Details/ Messages item',
   component: ApiRuntimeLogsMessageItemComponent,
   decorators: [
     moduleMetadata({
